+++ docs/src/examples/QasCard/Basic.vue
@@ -31,18 +31,18 @@ export default {
   computed: {
     images () {
       return [
-        'https://via.placeholder.com/200x300/ff0000/969696',
-        'https://via.placeholder.com/200x300/09ff00/969696',
-        'https://via.placeholder.com/200x300/0044ff/969696',
-        'https://via.placeholder.com/200x300/ff0000/969696',
-        'https://via.placeholder.com/200x300/09ff00/969696',
-        'https://via.placeholder.com/200x300/0044ff/969696',
-        'https://via.placeholder.com/200x300/ff0000/969696',
-        'https://via.placeholder.com/200x300/09ff00/969696',
-        'https://via.placeholder.com/200x300/0044ff/969696',
-        'https://via.placeholder.com/200x300/ff0000/969696',
-        'https://via.placeholder.com/200x300/09ff00/969696',
-        'https://via.placeholder.com/200x300/0044ff/969696'
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300',
+        'https://via.placeholder.com/200x300'
       ]
     }
   }