+++ ui/src/components/copy/QasCopy.vue
@@ -2,7 +2,7 @@
   <span>
     <slot>{{ text }}</slot>
 
-    <qas-btn class="q-ml-xs" color="grey-7" flat :icon="icon" :loading="isLoading" round :size="size" @click.stop="copy">
+    <qas-btn class="q-ml-xs" color="primary" :icon="icon" :loading="isLoading" variant="tertiary" @click.stop="copy">
       <q-tooltip>Copiar</q-tooltip>
     </qas-btn>
   </span>
@@ -17,11 +17,6 @@ export default {
   props: {
     icon: {
       default: 'sym_r_file_copy',
-      type: String
-    },
-
-    size: {
-      default: 'xs',
       type: String
     },